Fixed player_accuracy function.
git-svn-id: http://hopmod.googlecode.com/svn/trunk@75 d7818858-e1b4-11dd-a738-5fe2e3cfbcc5

+++ src/fpsgame/fpsserver.h
@@ -3092,7 +3092,7 @@ struct fpsserver : igameserver
         int hits=get_player_hits(cn);
         int misses=get_player_misses(cn);
         std::ostringstream out;
-        if(hits) out<<(int)((hits/(hits+misses))*100)<<"%";
+        if(hits) out<<(int)(((float)hits/(hits+misses))*100)<<"%";
         else out<<"0%";
         return out.str();
     }
